requirements.txt
streamlit
transformers
torch
langchain
langchain-community
sentence-transformers
unstructured
openai
tiktoken
groq
python-dotenv
python-dotenv
faiss-cpu
